+++ src/rendering/ansi.rs
@@ -1,5 +1,6 @@
 //! ANSI escape sequence related types and functions.
-use crate::style::color::Rgb;
+
+use embedded_graphics::pixelcolor::Rgb888;
 
 /// List of supported SGR (Select Graphics Rendition) sequences
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
@@ -20,19 +21,19 @@ pub enum Sgr {
     NotCrossedOut,
 
     /// Change the text color
-    ChangeTextColor(Rgb),
+    ChangeTextColor(Rgb888),
 
     /// Reset the text color to transparent
     DefaultTextColor,
 
     /// Change the background color
-    ChangeBackgroundColor(Rgb),
+    ChangeBackgroundColor(Rgb888),
 
     /// Reset the background color to transparent
     DefaultBackgroundColor,
 }
 
-fn try_parse_8b_color(v: &[u8]) -> Option<Rgb> {
+fn try_parse_8b_color(v: &[u8]) -> Option<Rgb888> {
     let color = *v.get(0)?;
     match color {
         //   0-  7:  standard colors (as in ESC [ 30–37 m)
@@ -53,50 +54,50 @@ fn try_parse_8b_color(v: &[u8]) -> Option<Rgb> {
             let (g, source_r) = extract_ch(source_rg);
             let (r, _) = extract_ch(source_r);
 
-            Some(Rgb::new(r, g, b))
+            Some(Rgb888::new(r, g, b))
         }
 
         // 232-255:  grayscale from black to white in 24 steps
         232..=255 => {
             let level = color - 232;
             let g = if level == 23 { 255 } else { level * 11 };
-            Some(Rgb::new(g, g, g))
+            Some(Rgb888::new(g, g, g))
         }
     }
 }
 
-fn try_parse_rgb(v: &[u8]) -> Option<Rgb> {
+fn try_parse_rgb(v: &[u8]) -> Option<Rgb888> {
     let r = *v.get(0)?;
     let g = *v.get(1)?;
     let b = *v.get(2)?;
 
-    Some(Rgb::new(r, g, b))
+    Some(Rgb888::new(r, g, b))
 }
 
-fn standard_to_rgb(idx: u8) -> Rgb {
+fn standard_to_rgb(idx: u8) -> Rgb888 {
     // These colors are used in PowerShell 6 in Windows 10
     match idx {
-        0 => Rgb::new(12, 12, 12),
-        1 => Rgb::new(197, 15, 31),
-        2 => Rgb::new(19, 161, 14),
-        3 => Rgb::new(193, 156, 0),
-        4 => Rgb::new(0, 55, 218),
-        5 => Rgb::new(136, 23, 152),
-        6 => Rgb::new(58, 150, 221),
-        7 => Rgb::new(204, 204, 204),
+        0 => Rgb888::new(12, 12, 12),
+        1 => Rgb888::new(197, 15, 31),
+        2 => Rgb888::new(19, 161, 14),
+        3 => Rgb888::new(193, 156, 0),
+        4 => Rgb888::new(0, 55, 218),
+        5 => Rgb888::new(136, 23, 152),
+        6 => Rgb888::new(58, 150, 221),
+        7 => Rgb888::new(204, 204, 204),
 
-        8 => Rgb::new(118, 118, 118),
-        9 => Rgb::new(231, 72, 86),
-        10 => Rgb::new(22, 198, 12),
-        11 => Rgb::new(249, 241, 165),
-        12 => Rgb::new(59, 120, 255),
-        13 => Rgb::new(180, 0, 158),
-        14 => Rgb::new(97, 214, 214),
-        _ => Rgb::new(242, 242, 242),
+        8 => Rgb888::new(118, 118, 118),
+        9 => Rgb888::new(231, 72, 86),
+        10 => Rgb888::new(22, 198, 12),
+        11 => Rgb888::new(249, 241, 165),
+        12 => Rgb888::new(59, 120, 255),
+        13 => Rgb888::new(180, 0, 158),
+        14 => Rgb888::new(97, 214, 214),
+        _ => Rgb888::new(242, 242, 242),
     }
 }
 
-fn try_parse_color(v: &[u8]) -> Option<Rgb> {
+fn try_parse_color(v: &[u8]) -> Option<Rgb888> {
     let color_type = *v.get(0)?;
 
     match color_type {